Frontend v3

+++ ZavrsnaAPP/ZavrsnaAPP/Models/Oruzje.cs
@@ -4,7 +4,7 @@
     {
         public string Naziv { get; set; }
         public string? Kalibar { get; set; }
-        public float? Cijena { get; set; }
+        public decimal? Cijena { get; set; }
         public int? Tezina { get; set; }
         public int Proizvodac { get; set; }
         public int? Kapacitetspremnika { get; set; }
